--- src/main/java/dev/heimz/core/definition/ModelDefinition.java
@@ -1,6 +1,0 @@
-package dev.heimz.core.definition;
-
-import org.immutables.value.Value.Immutable;
-
-@Immutable
-public interface ModelDefinition extends PolicyDefinition {}
