+++ src/main/java/dev/heimz/heimdall/definition/ModelDefinition.java
@@ -1,0 +1,8 @@
+package dev.heimz.heimdall.definition;
+
+import org.immutables.value.Value.Immutable;
+
+@Immutable
+public interface ModelDefinition extends PolicyDefinition {
+
+}
