+++ src/main/java/dev/heimz/heimdall/definition/PolicyDefinition.java
@@ -1,0 +1,22 @@
+package dev.heimz.heimdall.definition;
+
+import org.immutables.value.Value.Default;
+
+interface PolicyDefinition extends RoleDefinition, SubjectDefinition, RuleDefinition {
+
+    @Default
+    default boolean object() {
+        return true;
+    }
+
+    @Default
+    default boolean action() {
+        return true;
+    }
+
+    @Default
+    default boolean priority() {
+        return false;
+    }
+
+}
