+++ src/main/java/dev/heimz/heimdall/definition/SubjectDefinition.java
@@ -1,0 +1,22 @@
+package dev.heimz.heimdall.definition;
+
+import org.immutables.value.Value.Default;
+
+interface SubjectDefinition {
+
+    @Default
+    default boolean user() {
+        return true;
+    }
+
+    @Default
+    default boolean organization() {
+        return false;
+    }
+
+    @Default
+    default boolean group() {
+        return false;
+    }
+
+}
