+++ src/main/java/dev/heimz/heimdall/policy/Rule.java
@@ -1,0 +1,8 @@
+package dev.heimz.heimdall.policy;
+
+public enum Rule {
+    PERMIT,
+    RECOMMEND,
+    OBLIGE,
+    PROHIBIT
+}
